+++ fabfile/__init__.py
@@ -7,3 +7,4 @@ import es  # noqa
 import migrate  # noqa
 
 import v_0_1_0  # noqa
+import v_0_1_1  # noqa